+++ platform/jvm/capture/src/main/kotlin/io/bitdrift/capture/reports/IssueReporter.kt
@@ -162,19 +162,9 @@ internal class IssueReporter(
         if (Build.VERSION.SDK_INT < Build.VERSION_CODES.R) {
             return
         }
-        val lastReasonResult = latestAppExitInfoProvider.get()
-        if (lastReasonResult is LatestAppExitReasonResult.Valid) {
-            val lastReason = lastReasonResult.applicationExitInfo
-            lastReason.traceInputStream?.use {
-                latestAppExitInfoProvider.convertExitReasonToFbsReportType(lastReason.reason)?.let { fatalIssueType ->
-                    issueReporterProcessor?.processAppExitReport(
-                        fatalIssueType = fatalIssueType,
-                        timestamp = lastReason.timestamp,
-                        description = lastReason.description,
-                        traceInputStream = it,
-                    )
-                }
-            }
+        val latestAppExitReasonResult = latestAppExitInfoProvider.get()
+        if (latestAppExitReasonResult is LatestAppExitReasonResult.Valid) {
+            issueReporterProcessor?.processAppExitReport(latestAppExitReasonResult.applicationExitInfo)
         }
     }