+++ CHANGELOG.md
@@ -21,7 +21,7 @@
 
 **Added**
 
-- Nothing yet!
+- Native crash reports are now generated on API level 30 when tombstone data is unavailable. These reports include signal information but no stack traces.
 
 **Changed**